Refactor to use the download functionality used in the results section to download jobs on the multiple match page
