+++ src/aims_ui/static/js/b_multiple_matches/multiple_address_results/download_and_process/file_processing.mjs
@@ -104,7 +104,7 @@ function arrayToCSV(data) {
   return csvRows.join('\r\n') + returnNewlineIfNotBlank(csvRows);
 }
 
-export function processRow(row,page) {
+export function processRow(row) {
   // Process [id, inputAddress, APIresponse]
   // to be in same format as <5k match
