+++ src/aims_ui/static/js/f_helpers/download_helpers.mjs
